Update homework 1.py

+++ homework 1.py
@@ -1,6 +1,6 @@
 #homework 1
-a = int(input(" Enter the price of the land"))
-b = int(input(" Enter the price of the building"))
+a = float(input(" Enter the price of the land"))
+b = float(input(" Enter the price of the building"))
 c = float(1.21)
 total_cost_of_the_project = (a + b)*c
 print(f"The total cost of the project is: {total_cost_of_the_project}")
